Fixed winCondition, client-Server communication works till end turn
Win Condition: We now send the updated gameboard before declaring the winner.
Added starting cell parameter in RuleSetBase, we update the startingCell parameter once the move is validated

Added sout in player_move and player_build case in the client side parser.

Added selectedWorker parameter in SelectWorkerResponse message to update the selectedWorker in the client.
Client can select other worker

Removed BackReference tag in Cell-Worker EVERYTHING WORKS I DON'T KNOW WHY

Because the cell and worker that come from socket aren't the ones in the game's gameboard, in game.validateAction() we create another Action with the correct parameters. (we do the same thing in player.setSelectedWorker())