Implemented Visitor pattern for messages.
Added 2 abstract classes:
MessageFromServerToClient inherits Message and is inherited by all the messages sent by the server, all those messages implement callVisitor() that calls the correct method of the client Parser.
MessageFromClientToServer works in the same way but for messages sent by the client.

Added 2 visitor interfaces, one for the client Parser and one for the server parser, every switch case is now a method called by the message.