Modified god Json format; Implemented game reload from file
Modified JacksonTags in strategies to print a more readable json.

Added 2 messages to implement persistence;

Removed SaveState() from game; now the game is saved when a listener is notified inside serverController;
Modified restoreState() in game, now it doesn't deserialize the file but it restores the references between cells and workers, for some reason even if the json is serialized correctly the deserialization fails to keep bidirectional relationships intact.

Now when all the users are logged in the Lobby checks if a file named like any combination of the users name exists, if it exists it deserializes the file and calls restoreState() on the deserialized game.