NEEDS TESTING: Fixed game reloading, added settings file prototype
Now if the player selected a worker before the crash the game restarts from the choice of the action to do among the possible actions

Now if a config.txt file is found and is not empty the view will prompt a new message where the player will be able to reload one of his previous settings. The settings are stored from the last used to the eldest without repetition and they are saved once the server responds with a OK LoginResponse Message.
We store MAX_SETTINGS_STORED number of old settings