Modified client behaviour when connection drops
Extracted method from Client main, now we have a initClient that is called from the main and from the networkHandler when the connection drops.
As of now InitClient creates a new object Client everytime it is called, we should check if the old Client object remains Instantiated or if the old thread keeps running in the background.

Changed some sout in network handler and some Exception throws in order to have a more clear vision of what made the connection drop.

Removed a scanner in CLI showErrorMessage()