Fixed Prometheus "kill Himself" check and added tests, updated godJson
BUG: sometimes the client crashes while updating the cachedBoard (index out of bounds)

Added some checks in BuildBeforeAfterMovement in order to count how many stuck workers there are.
now if both workers are stuck the player will be able to build and then he will lose.
If a worker is able to play some legal actions without losing the player is forced to do one of those actions

updated GodJson to include stuckWorkers field in Prometheus.