Refactoring messages, some messages are managed asynchronously
NEED TO REMOVE SYNCHRONIZED IN VIRTUAL CLIENT - NEED TO FIX BUG WHILE CREATING LOBBY.
Added a boolean method in messages that indicates if a message has to be handled immediately.

Better managed some exceptions.

Removed refresh lobbies option from CLI, now the lobbies are refreshed as soon as a new one is created.

Modified the message constructors  in order to obtain the real time refresh lobby.

Added a method in server that sends a message to all users in the waiting room.