Added remaining input managers
Added isWaitingForInput setter in order to awake an input manager that could have finished, for example select worker has to stop waiting for input after selecting a cell, but if the user chose the wrong worker the input manager should start waiting for input again.

Move and Build actions are managed by the same class, the object know which one it has to manage thanks to a boolean in the constructor. (similar to lobbyInputManager).

Classes that performed completely different functions only based by the constructor used are now separated in multiple classes, this bring us to have a lot more classes but better designed.

I decided to not give every information possible to the input managers, for example managers never have access to the gameboard, selected worker and selected cell, those information are stored in the message parser; To create a move action the input manager has to call the method sendMove(Cell cell) of parser instead of sending the message directly.

Added a new Thread in selectAction because in case the player wanted to select the other worker we would have had a deadlock wile trying to do view.setInputManager().